Compare and align daily progress (#5)

+++ Hub/Hub.Application/Features/PlayerFeatures/Queries/GetPlayerProgress/GetPlayerProgressResponse.cs
@@ -2,5 +2,5 @@
 
 public class GetPlayerProgressResponse
 {
-    public Dictionary<string, decimal>? Progress { get; set; }
+    public Dictionary<string, int>? Progress { get; set; }
 }
